chore: move Selected Element styles to own file
