+++ app/common/renderer/components/SessionInspector/SourceTab/SelectedElement/InteractionsNotAvailableMessage.jsx
@@ -2,7 +2,7 @@ import {Alert, Col, Row} from 'antd';
 import {useTranslation} from 'react-i18next';
 
 import {ALERT, ROW} from '../../../../constants/antd-types.js';
-import styles from '../Source.module.css';
+import styles from './SelectedElement.module.css';
 
 /**
  * Info message shown when the selected element does not support interactions,
