+++ app/common/renderer/components/SessionInspector/SourceTab/SelectedElement/SelectedElement.module.css
@@ -1,0 +1,45 @@
+.selectedElementCard {
+  display: flex;
+  flex-direction: column;
+}
+
+.selectedElementCard :global(.ant-card-body) {
+  flex: 1;
+  overflow: auto;
+}
+
+.selectedElemInfoMessage {
+  justify-content: center;
+}
+
+.selectedElemActions {
+  gap: 8px;
+}
+
+.elementKeyInputActions {
+  max-width: 400px;
+  flex: 1;
+}
+
+.elementKeyInput {
+  height: 32px;
+  width: calc(100% - 64px);
+}
+
+.selectedElemTableWrapper {
+  display: flex;
+  width: 100%;
+}
+
+.selectedElemTableWrapper > div {
+  width: 100%;
+}
+
+.selectedElemTableCells {
+  word-wrap: break-all;
+  word-break: break-all;
+}
+
+.copyableCell {
+  cursor: pointer;
+}
