+++ app/common/renderer/components/SessionInspector/SourceTab/Source.module.css
@@ -1,49 +1,3 @@
-.selectedElementCard {
-  display: flex;
-  flex-direction: column;
-}
-
-.selectedElementCard :global(.ant-card-body) {
-  flex: 1;
-  overflow: auto;
-}
-
-.selectedElemInfoMessage {
-  justify-content: center;
-}
-
-.selectedElemTableCells {
-  word-wrap: break-all;
-  word-break: break-all;
-}
-
-.copyableCell {
-  cursor: pointer;
-}
-
-.selectedElemActions {
-  gap: 8px;
-}
-
-.elementKeyInputActions {
-  max-width: 400px;
-  flex: 1;
-}
-
-.elementKeyInput {
-  height: 32px;
-  width: calc(100% - 64px);
-}
-
-.selectedElemTableWrapper {
-  display: flex;
-  width: 100%;
-}
-
-.selectedElemTableWrapper > div {
-  width: 100%;
-}
-
 .sourceSearchHighlight {
   background-color: #ffeb3b;
   border-radius: 2px;
